server/code/requirements.txt
psycopg2-binary
git+https://github.com/meilof/pysnark
